Integrate Freq_output class into Crystal_output class - unfinished
